Java coding exercise: Randomizer with input
I created a randomizer that will take in names or numbers
entered by the user and saves them into an array. The script then
returns a random index from the array and prints it to the console.

Once it has printed out an index of the array it will delete that
index and repeat the proces until the array is empty.

It will also group the indexes into groups of 4 and print a
new line after each group of 4 indexes.